Resolve new warnings in UnitTests.

+++ Frameworks/AIUtilities Framework/Source/AIStringAdditions.h
@@ -38,7 +38,7 @@
 
 - (NSString *)safeFilenameString;
 
-- (NSString *)stringWithEllipsisByTruncatingToLength:(unsigned int)length;
+- (NSString *)stringWithEllipsisByTruncatingToLength:(NSUInteger)length;
 
 - (NSString *)string;
